{{cookiecutter.repo_slug}}/pyproject.toml
[project]
name = "{{cookiecutter.repo_slug}}"
version = "0.1.0"
requires-python = "=={{cookiecutter.python_version}}.*"
description = "Data engineering repo with Airflow (containers), Postgres, uv, ruff, sqlmodel, and dbt."
dependencies = [
  # Core data engineering stack
  "sqlmodel>=0.0.21",
  "pydantic>=2.7",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=2.9",
  "dbt-core>=1.7",
  
  # Configuration management
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  
  # Data processing
  "pandas>=2.0.0",
  "polars>=0.20.0",
  
  # Utilities
  "rich>=13.7.0",  # Better console output
  "typer>=0.9.0",  # CLI framework
]

[tool.uv]
# Use uv for resolver/installer

[tool.ruff]
line-length = 100
target-version = "py{{cookiecutter.python_version | replace('.', '')}}"
extend-select = ["I", "E", "F", "UP"]
fix = true
unsafe-fixes = true

[tool.ruff.lint.isort]
known-first-party = ["{{cookiecutter.repo_slug}}"]
combine-as-imports = true

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
